Fix OAuth proxy override typing
